chore(coordinator): split observability.rs into focused modules
observability.rs had grown to ~1460 lines across three concerns: the
metric registry with its derived series, the readiness and state-store
access wrapper, and the admin HTTP surface. Split along that seam into
metrics.rs, state.rs and admin.rs beneath observability/, with the
shared helpers and the re-export set in mod.rs.

Bodies move verbatim; the module boundary forces only pub(crate)
visibility on the internal members the siblings call, so the public
API is unchanged.

Refs #306
